Fix site restart on version update
Instead of using the init script provided by Check_MK use `omd` commands
for starting/stopping a site. This fixes the following two issues:

- If a version update was performed on one site, all other sites with the
same version on the server were stopped too (and never started again).

- After the update the startup was tried via Ansible 'service' modules
`state: started`. However, the service was already started after the
installation of the package just that it didn't had a site for this
version yet. This resulted in the site not being started anymore.